+++ src/docs/HISTORY.md
@@ -8,9 +8,11 @@ v1.7.6
 - UI: Added pencil annotations and erasing without a Solo License.  Note that area tools and other drawing tools like arrows remain on the Solo License.  This is a compromise between features present on all viewers and features only on mrv2/vmrv2 so that those that unlocked a Solo License, won't feel cheated.
 - UI: Added a Notes Panel to display and edit the list of review notes easily, as suggested by Serban Jipa.
 
+- Core: Added HW decoding and encoding on aarch64 platforms.
 - Core: Added support for OpenEXR's colorInteropID with the new Auto ICS button and its tooltip describing what it chose.  Auto ICS works with image files.  For videos, you should rely on **Preferences->OCIO->No OCIO on Videos or sRGB/BT709** data as it is faster.
 - Core: Added SVG reader.  Works with all .svg with paths.  Does not currently decode SVGs with text yet, though.
 - Core: Added HW encoding with NVidia (Windows and Linux), Intel/AMD (Linux) and macOS (was already present).  As usual, the features depend on your hardware and drivers.
+- Core: Added Vulkan HW decoding on Windows too.
 - Core: Added the NVidia headers to the build, so the Linux branch will compile without the need for local files.
 - Core: Added OpenColorIO's LUT inversion on Color Panel.